Change everything
This crate had a lot to answer for after its main purpose (the Contains trait) got added to the standard library. I was going to decommission it until I realised that it still has the Result type (and the generic type stuff) to offer.

So the Bounded<T> type has been replaced by the built-in RangeBounds<T> type, though we’ve had to keep the Bounds<T> type.